Added email confirmations to the web server
