+++ Project2/Project2/Project2.Android/Project2.Android.csproj.user
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
-    <SelectedDevice>android_nexus_nougat</SelectedDevice>
-    <DefaultDevice>android_nexus_nougat</DefaultDevice>
+    <SelectedDevice>Nexus_5X_API_16</SelectedDevice>
+    <DefaultDevice>Nexus_5X_API_16</DefaultDevice>
   </PropertyGroup>
 </Project>
